Simplifying the hit extraction

+++ include/fdreadoutlibs/wibeth/tpg/ProcessNaiveRS.hpp
@@ -62,8 +62,8 @@ process_window_naive_RS(ProcessingInfo<NREGISTERS>& info)
 
     // Variables for hit finding
     uint16_t& prev_was_over = state.prev_was_over[ichan]; // was the previous sample over threshold?
-    uint16_t& hit_charge = state.hit_charge[ichan];
-    uint16_t& hit_tover = state.hit_tover[ichan]; // time over threshold
+    uint32_t& hit_charge = state.hit_charge[ichan];
+    uint32_t& hit_tover = state.hit_tover[ichan]; // time over threshold
 
     for (size_t itime = 0; itime < info.timeWindowNumFrames; ++itime) {
       const size_t msg_index = itime / swtpg_wibeth::FRAMES_PER_MSG;
